requirements.txt
pyTelegramBotAPI>=4.12.0
requests>=2.31.0
feedparser>=6.0.10
APScheduler>=3.10.1
python-dotenv>=1.0.0
tenacity>=8.2.3 